requirements.txt
adbutils==1.2.11
androguard==3.3.5
apkutils2==1.0.0
appdirs==1.4.4
Appium-Python-Client==0.41
asn1crypto==1.5.1
backcall==0.2.0
bcrypt==4.0.1
cached-property==1.5.2
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
cigam==0.0.3
click==8.0.4
colorama==0.4.5
cryptography==3.4.8
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
Deprecated==1.2.14
deprecation==2.1.0
dnspython==2.2.1
et-xmlfile==1.1.0
filelock==3.4.1
future==0.18.3
futures==2.2.0
graphviz==0.19.1
idna==3.4
importlib-metadata==4.8.3
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
joblib==1.1.1
kiwisolver==1.3.1
logzero==1.7.0
lxml==4.9.3
matplotlib==3.3.4
mock==5.1.0
networkx==2.5.1
numpy==1.19.5
openpyxl==3.1.2
packaging==21.3
pandas==1.1.5
paramiko==3.4.0
parso==0.7.1
pickleshare==0.7.5
Pillow==8.4.0
progress==1.6
prompt-toolkit==3.0.36
psutil==5.9.7
py==1.11.0
pycparser==2.21
pycryptodome==3.20.0
pydot==1.4.2
pydotplus==2.0.2
pyelftools==0.30
Pygments==2.14.0
PyNaCl==1.5.0
pyparsing==3.1.1
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyshark==0.6
python-dateutil==2.8.2
python-magic==0.4.27
python-whois==0.8.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.15.1
retry==0.9.2
scikit-learn==0.24.2
scipy==1.5.4
selenium==4.0.0a3
six==1.16.0
termcolor==1.1.0
threadpoolctl==3.1.0
tomorrow==0.2.4
traitlets==4.3.3
typing_extensions==4.1.1
uiautomator2==2.16.25
urllib3==1.26.16
wcwidth==0.2.6
whichcraft==0.6.1
whois==0.9.27
wincertstore==0.2
wrapt==1.16.0
xmltodict==0.13.0
zipp==3.6.0
